Fix backups for tasks using a mounted file share
Currently, backing up such a task succeeds, but restoring fails with "No
such file or directory: <root>/share/manifest.jsonl".

IMO, the problem is not really in the restore logic, but in the backup
logic. Share task backups are "heavyweight", in that they include a copy of
the media files. But the `storage` field in `task.json` is still set to
"share". As a result, CVAT tries to import it as a share task, and that
logic is broken.

I think we should handle such backups consistently with heavyweight cloud
storage backups, and set `storage` to "local". The restore logic will then
work perfectly well. In the future, if we happen to implement lightweight
backups for share tasks, we can use `storage: "share"` for those.