relaxing manifest creation (#9994)
After update of ffmpeg, for some videos, when seeking to a keyframe,
seek lands on a previous keyframe.

With our current way of creating manifest, for such videos the manifest
is empty.
It is not optimal: it is a lot better to be able to seek to 1000th
keyframe of some video and land on 999th keyframe and decode to 1000th,
than to decode all the way from video beginning.

Relaxing manifest creation for such cases.