cvat-cli, utils: use PEP 604 syntax for optionals/unions (#10060)
Now that we dropped Python 3.9, we can use this, which is a little more
readable than `Union[X, Y]` and `Optional[X]`.

The following command was used to perform the replacement:

    ruff check --select=UP007,UP045 cvat-cli/ utils/ --fix

Since this change creates a fair bit of churn, I'm doing it piecemeal
rather than updating the whole codebase at once.