Fix task creation from cloud storage with static chunks and custom range
Currently this doesn't work, because:

1. `create_thread` calls `_download_data_from_cloud_storage` to download only
   those frames that match the frame range;

2. It tries to create a manifest, passing paths to _all_ frames.

3. `ImageManifestManager` tries to open a frame that hasn't been downloaded,
   and crashes.

Fix this by removing the frame filtering when downloading the frames. This
is obviously not optimal performance-wise, but it's better than not working
at all.

To be fair, this also slows down the case where a cloud storage manifest is
used. IMO, it is acceptable to take a performance hit in this case to reduce
complexity.