Fix heavyweight backups of tasks that have images and a non-default frame range (#10004)
Let's say we have a cloud storage-based task that was created with 6
input images, and the following settings: start frame 1, stop frame 4,
frame step 2. A heavyweight backup of such a task will then contain only
frames 1 and 3 in the `data` directory; however, the manifest will still
contain all 6 frames. If you try to restore such a backup, CVAT will
fail, because it checks that the files in the `data` directory
correspond 1:1 to the manifest.

This could potentially be fixed in the restore code; however, it seems
to me that the backups created in this case are incorrect, as they have
manifests referencing nonexistent files. As such, I think it's more
appropriate to fix this in the backup code.

The fix is to filter the manifest during backup, leaving only entries
corresponding to frames that actually get backed up. We also need to
reset the frame range to the default, so that it matches the filtered
manifest.

Note that the same bug also affects backups of tasks based on the
mounted share. It should be reasonably easy to fix (just use
`_write_filtered_media_manifest` in the `StorageChoice.SHARE` branch),
but I cannot test such a fix, because share backups are currently broken
entirely. So I will defer this fix until #9972.

This bug does _not_ affect backups of tasks based on local files,
because those currently include _all_ frames, even those that fall
outside that frame range. I think that behavior is probably wrong, but
it doesn't require immediate fixing.