Fix backups for tasks using a mounted file share (#9972)
Fixes #4989
Fixes #6149
Fixes #7965
Fixes #8297

There are three problems here.

First, CVAT refuses to back up such tasks if they use static chunks.
There doesn't seem to be much purpose for this; the backup process
doesn't even use chunks. With the check removed, the backups work fine.

Second, even when backing up succeeds, restoring fails with "No such
file or directory: <root>/share/manifest.jsonl".

IMO, the problem here is not really in the restore logic, but in the
backup logic. Share task backups are "heavyweight", in that they include
a copy of the media files. But the `storage` field in `task.json` is
still set to "share". As a result, CVAT tries to import it as a share
task, and that logic is broken.

I think we should handle such backups consistently with heavyweight
cloud storage backups, and set `storage` to "local". The restore logic
will then work perfectly well. In the future, if we happen to implement
lightweight backups for share tasks, we can use `storage: "share"` for
those.

Third, if a task has images and a custom frame range (start frame/stop
frame/step), then importing the backup _also_ doesn't work for reasons
described in #10004. We can fix this using the
`_write_filtered_media_manifest` function introduced in that PR.

This PR also removes logic from `create_thread` that tries to fix up
backups that have `storage="share"`. After this patch, such backups no
longer need fixing up (and before the patch, they couldn't be imported
anyway).