RAM optimization for sparse tracks on export (#10041)
When tracks are exported (in most formats), interpolation happens and
interpolated shapes are kept in RAM.
E.g. if a track is 100 frames long and has 10 keyframes, only 10 shapes
are read from db and kept in RAM before the interpolation.
Then, after interpolation, there are 100 shapes which are kept in RAM
during export - 10x more ram is used.

This PR makes interpolation function into a generator, to not keep all
those non-key shapes in RAM.