Use `container.decode` instead of `container.demux`
This is a helper method that pretty much just wraps the double loop over the
packets and the frames in a packet. Using it makes the code simpler and
enables better type inference in IDEs, because the PyAV type stubs are
sufficient to infer that `container.decode(video_stream)` returns an
`Iterable[VideoFrame]` (whereas `packet.decode()` is declared to return
`list[SubtitleSet]` for some reason).

Also, add some imports and type annotations to ensure that the IDEs have types
to infer _from_.