Merge commit from fork
It is trivial for an attacker to pass a directory like `../../../../` and
view the contents of any directory on the server.

There is code to prevent this, but it doesn't work:
`(Path(settings.SHARE_ROOT) / directory_param).absolute()` just returns a
path like `/home/django/share/../../../../`, which starts with
`/home/django/share`, so it passes the check.

Fix it by using `.resolve()` instead.