`media_extractors.py`: end plural confusion
Currently, we have variables named `source_path` that contain lists of
multiple paths and `_source_paths` that sometimes contain lists of a single
element. This is rather confusing. Rename and refactor everything so that
singular nouns correspond to single paths, and plural nouns to lists.

Note that the extractor constructors still have to take lists even when they
only expect a single element, in order to conform to the general interface.
Use list unpacking in such constructors; this ensures that they will crash
if used incorrectly.

Also, rename the `reconcile` argument from `source_files` to `source_paths`
for consistency.