Fix video manifests getting ignored during task creation (#10162)
`VideoManifestValidator` expects an `Openable` as the `source_path`, but
here a string is passed. This would normally crash, but it's inside a
`try...except` block that suppresses all exceptions, so what actually
happens is that every time a video is uploaded together with a manifest,
that manifest is ignored and CVAT tries to generate a new one.

This error handling should probably be improved, but for now, just
restore intended functionality.